third_party/blink/web_tests/fast/forms/radio/radio-group-arrow-cycle-edge-expected.txt

Tests after pressing last radio button in the list, on pressing downkey it selects back first radio button

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("cherry").checked is true
PASS document.getElementById("apple").checked is true
PASS document.getElementById("cherry").checked is true
PASS document.getElementById("cucumber").checked is true
PASS document.getElementById("tomato").checked is true
PASS document.getElementById("cucumber").checked is true
PASS document.getElementById("pumpkin").checked is true
PASS document.getElementById("pumpkin").checked is true
PASS document.getElementById("pumpkin").checked is true
PASS successfullyParsed is true

TEST COMPLETE
